backend-futu/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
futu-api==9.6.5608
python-dotenv==1.0.0
